requirements.txt
langchain
tiktoken 
openai
pinecone-client
nltk
streamlit
streamlit-chat
tqdm
beautifulsoup4